requirements.txt
llama-index>=0.11.0
llama-index-vector-stores-chroma
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-readers-file
streamlit>=1.31.0
chromadb>=0.4.22
openai>=1.12.0
pillow>=10.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
pdfminer.six>=20221105
pdfplumber>=0.10.0
pymupdf>=1.23.0
